fixed broken test + made 1 file for the token sale example.
