+++ pkg/vm/tests/foobar/bar.go
@@ -1,0 +1,3 @@
+package foobar
+
+func getBool() bool { return true }
