+++ pkg/vm/tests/foobar/foo.go
@@ -1,0 +1,7 @@
+package foobar
+
+// OtherBool ...
+func OtherBool() bool {
+	ok := getBool()
+	return ok
+}
